Refacto(parser): Fix linter
