+++ include/parser.h
@@ -1,0 +1,45 @@
+/*
+** EPITECH PROJECT, 2023
+** 42sh_repo
+** File description:
+** No file there , just an epitech header example .
+*/
+
+#ifndef INC_42SH_REPO_PARSER_H
+    #define INC_42SH_REPO_PARSER_H
+
+    #include <stdlib.h>
+
+    #define TI token->index
+    #define TL token->line
+
+typedef struct pipe_s {
+    int *fds;
+    int max;
+    int index;
+} pipe_t;
+
+typedef struct token_s {
+    int index;
+    int type;
+    char *line;
+    char *elem;
+    int *exit;
+    int *error;
+    pipe_t *pipes;
+} token_t;
+
+typedef enum special_type_e {
+    END = -1,
+    NONE = 0,
+    REDIR_RIGHT,
+    REDIR_LEFT,
+    DOUBLE_REDIR_RIGHT,
+    DOUBLE_REDIR_LEFT,
+    PIPE,
+    SEMICOLON,
+    AND,
+    OR
+} special_type_e;
+
+#endif //INC_42SH_REPO_PARSER_H
